requirements.txt
black==24.3.0
flake8==6.0.0
isort==5.12.0
loguru==0.6.0
pip==25.1.1
pylint==3.0.3
pytest==7.3.1
pytest-cov==4.1.0
python-dotenv==1.0.0
tqdm==4.66.3
typer==0.7.0
Flask==2.3.3
flasgger==0.9.7.1
rouge_score==0.1.2
scikit-learn==1.6.1
pandas==2.2.0
pathlib==1.0.1
joblib==1.4.2
dvc[gdrive]
lib-ml-remla-team3 @ git+https://github.com/remla25-team3/lib-ml@v0.4.0

-e .